+++ lib/ruby_speech/ssml.rb
@@ -13,12 +13,10 @@ module RubySpeech
     InvalidChildError = Class.new StandardError
 
     def self.draw(&block)
-      Nokogiri::XML::Document.new.tap do |doc|
-        doc << Speak.new.tap do |speak|
-          block_return = speak.instance_eval(&block) if block_given?
-          speak << block_return if block_return.is_a?(String)
-        end
-      end.to_s
+      Speak.new.tap do |speak|
+        block_return = speak.instance_eval(&block) if block_given?
+        speak << block_return if block_return.is_a?(String)
+      end
     end
   end # SSML
 end # RubySpeech