+++ spec/ruby_speech/ssml/speak_spec.rb
@@ -118,6 +118,31 @@ module RubySpeech
           lambda { subject << 1 }.should raise_error(InvalidChildError, "A Speak can only accept String, Audio, Break, Emphasis, Mark, P, Phoneme, Prosody, SayAs, Sub, S, Voice as children")
         end
       end
+
+      describe "#to_doc" do
+        let(:expected_doc) do
+          Nokogiri::XML::Document.new.tap do |doc|
+            doc << Speak.new
+          end
+        end
+
+        it "should create an XML document from the speak" do
+          Speak.new.to_doc.to_s.should == expected_doc.to_s
+        end
+      end
+
+      it "should allow concatenation" do
+        speak1 = Speak.new
+        speak1 << Voice.new(name: 'frank', content: "Hi, I'm Frank")
+        speak2 = Speak.new
+        speak2 << Voice.new(name: 'millie', content: "Hi, I'm Millie")
+
+        expected_concat = Speak.new
+        expected_concat << Voice.new(name: 'frank', content: "Hi, I'm Frank")
+        expected_concat << Voice.new(name: 'millie', content: "Hi, I'm Millie")
+
+        (speak1 + speak2).to_s.should == expected_concat.to_s
+      end
     end # Speak
   end # SSML
 end # RubySpeech