+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.2
+  API Change: SSML.draw now returns a Speak
+  Feature: Speak objects can be turned into an XML document using #to_doc
+  Feature: Speak objects can now be concatenated such that children are merged together
+
 # 0.1.1
   * Bugfix: DSL now allows for nesting all allowed elements within each other